+++ lib/screens/login_screen.dart
@@ -2,9 +2,9 @@ import 'package:flutter/material.dart';
 import '../services/auth_service.dart';
 import 'conversations_screen.dart';
 import '../widgets/custom_icon.dart';
+import 'login_email_screen.dart';
 
-/// Tela de login do aplicativo
-/// Permite autenticação anônima, email/senha, Google e navegação para o menu de conversas
+/// Tela de login principal: limpa, só com botões
 class LoginScreen extends StatefulWidget {
   const LoginScreen({super.key});
 
@@ -14,22 +14,8 @@ class LoginScreen extends StatefulWidget {
 
 class _LoginScreenState extends State<LoginScreen> {
   final AuthService _authService = AuthService();
-  final _formKey = GlobalKey<FormState>();
-  final _emailController = TextEditingController();
-  final _passwordController = TextEditingController();
-  
   bool _isLoading = false;
-  bool _isLoginMode = true; // true = login, false = registro
-  bool _obscurePassword = true;
 
-  @override
-  void dispose() {
-    _emailController.dispose();
-    _passwordController.dispose();
-    super.dispose();
-  }
-
-  /// Realiza o login anônimo
   Future<void> _signInAnonymously() async {
     setState(() => _isLoading = true);
     try {
@@ -50,7 +36,6 @@ class _LoginScreenState extends State<LoginScreen> {
     }
   }
 
-  /// Realiza login com Google
   Future<void> _signInWithGoogle() async {
     setState(() => _isLoading = true);
     try {
@@ -71,71 +56,10 @@ class _LoginScreenState extends State<LoginScreen> {
     }
   }
 
-  /// Realiza login/registro com email e senha
-  Future<void> _signInWithEmailAndPassword() async {
-    if (!_formKey.currentState!.validate()) return;
-    
-    setState(() => _isLoading = true);
-    try {
-      if (_isLoginMode) {
-        await _authService.signInWithEmailAndPassword(
-          _emailController.text.trim(),
-          _passwordController.text,
-        );
-      } else {
-        await _authService.createUserWithEmailAndPassword(
-          _emailController.text.trim(),
-          _passwordController.text,
-        );
-      }
-      
-      if (mounted) {
-        Navigator.of(context).pushReplacement(
-          MaterialPageRoute(builder: (context) => const ConversationsScreen()),
-        );
-      }
-    } catch (e) {
-      if (mounted) {
-        final errorMessage = _authService.getErrorMessage(e.toString());
-        ScaffoldMessenger.of(context).showSnackBar(
-          SnackBar(content: Text(errorMessage), backgroundColor: Colors.red),
-        );
-      }
-    } finally {
-      if (mounted) setState(() => _isLoading = false);
-    }
-  }
-
-  /// Envia email de recuperação de senha
-  Future<void> _sendPasswordResetEmail() async {
-    if (_emailController.text.trim().isEmpty) {
-      ScaffoldMessenger.of(context).showSnackBar(
-        const SnackBar(content: Text('Digite seu email primeiro'), backgroundColor: Colors.orange),
-      );
-      return;
-    }
-
-    setState(() => _isLoading = true);
-    try {
-      await _authService.sendPasswordResetEmail(_emailController.text.trim());
-      if (mounted) {
-        ScaffoldMessenger.of(context).showSnackBar(
-          const SnackBar(
-            content: Text('Email de recuperação enviado!'),
-            backgroundColor: Colors.green,
-          ),
-        );
-      }
-    } catch (e) {
-      if (mounted) {
-        final errorMessage = _authService.getErrorMessage(e.toString());
-        ScaffoldMessenger.of(context).showSnackBar(
-          SnackBar(content: Text(errorMessage), backgroundColor: Colors.red),
-        );
-      }
-    } finally {
-      if (mounted) setState(() => _isLoading = false);
-    }
+  void _goToEmailLogin() {
+    Navigator.of(context).push(
+      MaterialPageRoute(builder: (context) => const LoginEmailScreen()),
+    );
   }
 
   @override
@@ -145,155 +69,77 @@ class _LoginScreenState extends State<LoginScreen> {
         title: const Text('Chat Simulator'),
         centerTitle: true,
       ),
-      body: SingleChildScrollView(
-        padding: const EdgeInsets.all(24),
+      body: Center(
         child: Column(
           mainAxisAlignment: MainAxisAlignment.center,
           children: [
-            const SizedBox(height: 40),
             const CustomIcon(
               emoji: '💬',
               size: 100,
               color: Colors.blue,
             ),
             const SizedBox(height: 32),
-            Text(
-              _isLoginMode ? 'Entrar' : 'Criar Conta',
-              style: const TextStyle(
-                fontSize: 28,
-                fontWeight: FontWeight.bold,
-              ),
+            const Text(
+              'Oi! Este é o Chat Simulator.',
+              style: TextStyle(fontSize: 28, fontWeight: FontWeight.bold),
             ),
             const SizedBox(height: 16),
             const Text(
               'Simule conversas totalmente controladas por você 😉',
-              style: TextStyle(
-                fontSize: 16,
-                color: Colors.grey,
-              ),
+              style: TextStyle(fontSize: 16, color: Colors.grey),
               textAlign: TextAlign.center,
             ),
-            const SizedBox(height: 32),
-            
-            // Formulário de email/senha
-            Form(
-              key: _formKey,
+            const SizedBox(height: 48),
+            // Botões centralizados, um abaixo do outro
+            SizedBox(
+              width: 260,
               child: Column(
+                crossAxisAlignment: CrossAxisAlignment.stretch,
                 children: [
-                  TextFormField(
-                    controller: _emailController,
-                    keyboardType: TextInputType.emailAddress,
-                    decoration: const InputDecoration(
-                      labelText: 'Email',
-                      prefixIcon: Icon(Icons.email),
-                      border: OutlineInputBorder(),
-                    ),
-                    validator: (value) {
-                      if (value == null || value.trim().isEmpty) {
-                        return 'Digite seu email';
-                      }
-                      if (!value.contains('@')) {
-                        return 'Digite um email válido';
-                      }
-                      return null;
-                    },
-                  ),
-                  const SizedBox(height: 16),
-                  TextFormField(
-                    controller: _passwordController,
-                    obscureText: _obscurePassword,
-                    decoration: InputDecoration(
-                      labelText: 'Senha',
-                      prefixIcon: const Icon(Icons.lock),
-                      suffixIcon: IconButton(
-                        icon: Icon(_obscurePassword ? Icons.visibility : Icons.visibility_off),
-                        onPressed: () => setState(() => _obscurePassword = !_obscurePassword),
-                      ),
-                      border: const OutlineInputBorder(),
-                    ),
-                    validator: (value) {
-                      if (value == null || value.isEmpty) {
-                        return 'Digite sua senha';
-                      }
-                      if (value.length < 6) {
-                        return 'A senha deve ter pelo menos 6 caracteres';
-                      }
-                      return null;
-                    },
-                  ),
-                  const SizedBox(height: 24),
-                  
-                  // Botão de login/registro
-                  SizedBox(
-                    width: double.infinity,
-                    child: ElevatedButton(
-                      onPressed: _isLoading ? null : _signInWithEmailAndPassword,
-                      style: ElevatedButton.styleFrom(
-                        padding: const EdgeInsets.symmetric(vertical: 16),
-                      ),
-                      child: _isLoading
-                          ? const SizedBox(
-                              height: 20,
-                              width: 20,
-                              child: CircularProgressIndicator(strokeWidth: 2),
-                            )
-                          : Text(_isLoginMode ? 'Entrar' : 'Criar Conta'),
-                    ),
-                  ),
-                  
-                  // Links de ação
-                  const SizedBox(height: 16),
-                  Row(
-                    mainAxisAlignment: MainAxisAlignment.spaceBetween,
-                    children: [
-                      TextButton(
-                        onPressed: _isLoading ? null : () {
-                          setState(() => _isLoginMode = !_isLoginMode);
-                        },
-                        child: Text(_isLoginMode ? 'Criar conta' : 'Já tenho conta'),
-                      ),
-                      if (_isLoginMode)
-                        TextButton(
-                          onPressed: _isLoading ? null : _sendPasswordResetEmail,
-                          child: const Text('Esqueci a senha'),
-                        ),
-                    ],
-                  ),
-                ],
-              ),
-            ),
-            
-            const SizedBox(height: 32),
-            const Text('ou', style: TextStyle(color: Colors.grey)),
-            const SizedBox(height: 24),
-            
-            // Botões de login alternativos
-            Row(
-              children: [
-                Expanded(
-                  child: ElevatedButton.icon(
+                  ElevatedButton.icon(
                     onPressed: _isLoading ? null : _signInAnonymously,
-                    icon: const CustomIcon(emoji: '👤', size: 20),
+                    icon: const CustomIcon(emoji: '👤', size: 22),
                     label: const Text('Anônimo'),
                     style: ElevatedButton.styleFrom(
                       padding: const EdgeInsets.symmetric(vertical: 12),
+                      textStyle: const TextStyle(fontSize: 16),
+                      shape: RoundedRectangleBorder(
+                        borderRadius: BorderRadius.circular(30),
+                      ),
                     ),
                   ),
-                ),
-                const SizedBox(width: 16),
-                Expanded(
-                  child: ElevatedButton.icon(
+                  const SizedBox(height: 16),
+                  ElevatedButton.icon(
                     onPressed: _isLoading ? null : _signInWithGoogle,
-                    icon: const CustomIcon(emoji: '🇬', size: 20),
+                    icon: const CustomIcon(emoji: '🇬', size: 22),
                     label: const Text('Google'),
                     style: ElevatedButton.styleFrom(
                       padding: const EdgeInsets.symmetric(vertical: 12),
+                      textStyle: const TextStyle(fontSize: 16),
                       backgroundColor: Colors.white,
                       foregroundColor: Colors.black,
+                      shape: RoundedRectangleBorder(
+                        borderRadius: BorderRadius.circular(30),
+                      ),
                     ),
                   ),
-                ),
-              ],
+                  const SizedBox(height: 16),
+                  ElevatedButton.icon(
+                    onPressed: _isLoading ? null : _goToEmailLogin,
+                    icon: const CustomIcon(emoji: '✉️', size: 22),
+                    label: const Text('Email'),
+                    style: ElevatedButton.styleFrom(
+                      padding: const EdgeInsets.symmetric(vertical: 12),
+                      textStyle: const TextStyle(fontSize: 16),
+                      backgroundColor: Colors.blue.shade50,
+                      foregroundColor: Colors.blue.shade900,
+                      shape: RoundedRectangleBorder(
+                        borderRadius: BorderRadius.circular(30),
+                      ),
+                    ),
+                  ),
+                ],
+              ),
             ),
           ],
         ),
